feat(footer-update): move social media icons to footer and refactor component usage
