+++ src/components/social-media-icon.tsx
@@ -1,21 +1,18 @@
 const SocialMediaIcon = ({
-  type,
   link,
   icon: Icon
 }: {
-  type: string
   link: string
   icon: React.ElementType
 }) => {
   return (
     <a
-      key={type}
       href={link}
       target="_blank"
       rel="noopener noreferrer"
-      className="p-2 transition-colors hover:text-teal-500 dark:hover:text-[#ff63c3]"
+      className="p-2 transition-colors text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white"
     >
-      <Icon className="w-10 h-10" />
+      <Icon className="w-6 h-6" />
     </a>
   )
 }
